Fixes and Cleanup
Fixed a lot of errors with the transfer from using PyQt5 to Tkinter and removed some old files that were specific to PyQt5. Also prepared the new UI for use, the buttons need to have a button_name_clicked() function for each one still, but that will be added later.