Finalized GUI Design
Finished the GUI design and the program is fully functional. Additional touches will be added next such as more detailed code comments and unittest or pytest will also be added next. Exception handling may need to be added, but since there is a limited number of user input it may not.